Add rendering settings files for 3 and 4-channel images
Looking at the plates, only the first 5 ones have data for 4 channels while
the following have only 3. This allows to adjust the rendering for the
reimported plates to match the names and colors used across the equivalent
plates in the screen.
Note this assumes the min/max values have been computed at import time.